Move AbstractConsoleGroupedTaskFunctionalTest to internalIntegTesting
